+++ src/main/java/com/lgcns/newspacebackend/global/security/config/CorsConfig.java
@@ -1,35 +1,35 @@
-package com.lgcns.newspacebackend.global.security.config;
-
-import java.util.List;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
-
-@Configuration
-public class CorsConfig {
-	@Bean
-    public CorsFilter corsFilter() {
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration config = new CorsConfiguration();
-        config.setAllowCredentials(true);
-        config.setAllowedOriginPatterns(List.of("*"));
-//		config.setAllowedOrigins(List.of(
-//                "http://kudong.kr:55020", 
-//                "http://kudong.kr:55021", 
-//                "http://kudong.kr:55022", 
-//                "http://kudong.kr:55023",  
-//                "http://kudong.kr:55025",  
-//                "http://localhost:8080",
-//                "http://localhost:8070",
-//                "http://localhost:5173"));
-        config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"));
-        config.addAllowedHeader("*");
-        config.addExposedHeader("*");
-        config.setMaxAge(3600L);
-        source.registerCorsConfiguration("/**", config);
-        return new CorsFilter(source);
-    }
-}
+//package com.lgcns.newspacebackend.global.security.config;
+//
+//import java.util.List;
+//
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.web.cors.CorsConfiguration;
+//import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+//import org.springframework.web.filter.CorsFilter;
+//
+//@Configuration
+//public class CorsConfig {
+//	@Bean
+//    public CorsFilter corsFilter() {
+//        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+//        CorsConfiguration config = new CorsConfiguration();
+//        config.setAllowCredentials(true);
+//        config.setAllowedOriginPatterns(List.of("*"));
+////		config.setAllowedOrigins(List.of(
+////                "http://kudong.kr:55020", 
+////                "http://kudong.kr:55021", 
+////                "http://kudong.kr:55022", 
+////                "http://kudong.kr:55023",  
+////                "http://kudong.kr:55025",  
+////                "http://localhost:8080",
+////                "http://localhost:8070",
+////                "http://localhost:5173"));
+//        config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"));
+//        config.addAllowedHeader("*");
+//        config.addExposedHeader("*");
+//        config.setMaxAge(3600L);
+//        source.registerCorsConfiguration("/**", config);
+//        return new CorsFilter(source);
+//    }
+//}
