+++ src/main/java/com/lgcns/newspacebackend/global/security/config/SecurityConfig.java
@@ -8,6 +8,8 @@ import com.lgcns.newspacebackend.global.security.filter.JwtAuthorizationFilter;
 import com.lgcns.newspacebackend.global.security.jwt.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
 
+import java.util.List;
+
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,7 +21,11 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.access.channel.ChannelProcessingFilter;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
 
 @Configuration
 @EnableWebSecurity
@@ -82,6 +88,28 @@ public class SecurityConfig{
 //        return http.build();
 //    }
     @Bean
+    public CorsFilter corsFilter() {
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowCredentials(true);
+        config.setAllowedOriginPatterns(List.of("*"));
+//		config.setAllowedOrigins(List.of(
+//                "http://kudong.kr:55020", 
+//                "http://kudong.kr:55021", 
+//                "http://kudong.kr:55022", 
+//                "http://kudong.kr:55023",  
+//                "http://kudong.kr:55025",  
+//                "http://localhost:8080",
+//                "http://localhost:8070",
+//                "http://localhost:5173"));
+        config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"));
+        config.addAllowedHeader("*");
+        config.addExposedHeader("*");
+        config.setMaxAge(3600L);
+        source.registerCorsConfiguration("/**", config);
+        return new CorsFilter(source);
+    }
+    @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http.cors();
     	http.csrf((csrf) -> csrf.disable());
@@ -102,6 +130,7 @@ public class SecurityConfig{
                         .anyRequest().permitAll()
         );
 
+        http.addFilterBefore(corsFilter(), ChannelProcessingFilter.class);
         // JWT 인증 및 인가 필터 추가
         http.addFilterBefore(jwtAuthorizationFilter(), JwtAuthenticationFilter.class);
         http.addFilterBefore(jwtAuthenticationFilter(userRepository), UsernamePasswordAuthenticationFilter.class);
